requirements.txt
fastapi==0.115.0
sqlalchemy==2.0.35
pydantic==2.10.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17